+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/npc/NPCs.kt
@@ -4,7 +4,6 @@ import com.github.michaelbull.logging.InlineLogger
 import world.gregs.voidps.engine.entity.*
 import world.gregs.voidps.engine.entity.character.CharacterList
 import world.gregs.voidps.engine.entity.definition.NPCDefinitions
-import world.gregs.voidps.engine.entity.list.MAX_NPCS
 import world.gregs.voidps.engine.event.EventHandlerStore
 import world.gregs.voidps.engine.map.Tile
 import world.gregs.voidps.engine.map.collision.CollisionStrategyProvider
