+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/Players.kt
@@ -1,7 +1,7 @@
 package world.gregs.voidps.engine.entity.character.player
 
+import world.gregs.voidps.engine.entity.MAX_PLAYERS
 import world.gregs.voidps.engine.entity.character.CharacterList
-import world.gregs.voidps.engine.entity.list.MAX_PLAYERS
 
 class Players : CharacterList<Player>(MAX_PLAYERS) {
