+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/gfx/Graphics.kt
@@ -1,7 +1,7 @@
 package world.gregs.voidps.engine.entity.gfx
 
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap
-import world.gregs.voidps.engine.entity.list.BatchList
+import world.gregs.voidps.engine.entity.BatchList
 
 class Graphics(
     override val chunks: MutableMap<Int, MutableList<AreaGraphic>> = Int2ObjectOpenHashMap()
