+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/proj/Projectiles.kt
@@ -1,7 +1,7 @@
 package world.gregs.voidps.engine.entity.proj
 
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap
-import world.gregs.voidps.engine.entity.list.BatchList
+import world.gregs.voidps.engine.entity.BatchList
 
 class Projectiles(
     override val chunks: MutableMap<Int, MutableList<Projectile>> = Int2ObjectOpenHashMap()
