+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/sound/Sounds.kt
@@ -1,7 +1,7 @@
 package world.gregs.voidps.engine.entity.sound
 
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap
-import world.gregs.voidps.engine.entity.list.BatchList
+import world.gregs.voidps.engine.entity.BatchList
 
 class Sounds(
     override val chunks: MutableMap<Int, MutableList<AreaSound>> = Int2ObjectOpenHashMap()
