+++ engine/src/test/kotlin/world/gregs/voidps/engine/client/update/npc/NPCMovementTaskTest.kt
@@ -14,8 +14,8 @@ import world.gregs.voidps.engine.entity.character.move.moving
 import world.gregs.voidps.engine.entity.character.move.running
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.character.npc.NPCs
+import world.gregs.voidps.engine.entity.entityListModule
 import world.gregs.voidps.engine.entity.hasEffect
-import world.gregs.voidps.engine.entity.list.entityListModule
 import world.gregs.voidps.engine.event.eventModule
 import world.gregs.voidps.engine.map.Delta
 import world.gregs.voidps.engine.map.collision.blocked
