+++ engine/src/test/kotlin/world/gregs/voidps/engine/client/update/player/PlayerUpdateTaskTest.kt
@@ -16,7 +16,7 @@ import world.gregs.voidps.engine.client.update.view.Viewport
 import world.gregs.voidps.engine.entity.Direction
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.Players
-import world.gregs.voidps.engine.entity.list.entityListModule
+import world.gregs.voidps.engine.entity.entityListModule
 import world.gregs.voidps.engine.event.eventModule
 import world.gregs.voidps.engine.map.Delta
 import world.gregs.voidps.engine.map.Tile
